1 E-Commerces/requirements.txt
flask
flask-cors
mysql-connector-python
pymongo
bcrypt
